requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.2
notion-client>=2.2.1
python-dotenv>=1.0.0
pandas>=2.0.0
aiohttp>=3.9.0